read/write return ssize_t
git-svn-id: https://svn.r-project.org/R/trunk@59108 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/parallel/src/fork.c
@@ -257,7 +257,7 @@ SEXP mc_send_master(SEXP what)
     }
     while (i < len)
     {
-        int n = write(master_fd, b + i, len - i);
+        ssize_t n = write(master_fd, b + i, len - i);
         if (n < 1)
         {
             close(master_fd);
@@ -292,7 +292,7 @@ SEXP mc_send_child_stdin(SEXP sPid, SEXP what)
     fd = ci->sifd;
     while (i < len)
     {
-        int n = write(fd, b + i, len - i);
+        ssize_t n = write(fd, b + i, len - i);
         if (n < 1)
             error(_("write error"));
         i += n;
@@ -434,7 +434,7 @@ static SEXP read_child_ci(child_info_t *ci)
 {
     unsigned int len = 0;
     int fd = ci->pfd;
-    int n = read(fd, &len, sizeof(len));
+    ssize_t n = read(fd, &len, sizeof(len));
 #ifdef MC_DEBUG
     Dprintf(" read_child_ci(%d) - read length returned %d\n", ci->pid, n);
 #endif
